+++ src/main/kotlin/online/bingzi/luck/perms/bridge/spring/boot/starter/event/model/MessageEvents.kt
@@ -1,15 +1,16 @@
 package online.bingzi.luck.perms.bridge.spring.boot.starter.event.model
 
-import online.bingzi.luck.perms.bridge.spring.boot.starter.event.AbstractLuckPermsEvent
 import online.bingzi.luck.perms.bridge.spring.boot.starter.event.EventPriority
 import online.bingzi.luck.perms.bridge.spring.boot.starter.event.EventType
+import online.bingzi.luck.perms.bridge.spring.boot.starter.event.LuckPermsEvent
 
 /**
  * 自定义消息事件
  * 用于处理自定义消息的接收
  */
-data class CustomMessageEvent(
+class CustomMessageEvent(
+    source: Any,
     val channel: String,
     val message: String,
-    override val priority: EventPriority = EventPriority.NORMAL
-) : AbstractLuckPermsEvent(EventType.CUSTOM_MESSAGE) 
+    priority: EventPriority = EventPriority.NORMAL
+) : LuckPermsEvent(source, EventType.CUSTOM_MESSAGE, priority)